+++ Patches/MainMenuPatch.cs
@@ -6,6 +6,14 @@ namespace PeakGeneralImprovements.Patches
 {
     internal static class MainMenuPatch
     {
+        [HarmonyPatch(typeof(MainMenu), nameof(Start))]
+        [HarmonyPostfix]
+        private static void Start()
+        {
+            // This has to be called after Steam starts up because it reads user stats. Run it each time the menu loads.
+            Plugin.CalculateQuickStartAscent();
+        }
+
         [HarmonyPatch(typeof(MainMenu), "PlaySoloClicked")]
         [HarmonyTranspiler]
         private static IEnumerable<CodeInstruction> PlaySoloClicked_Transpiler(IEnumerable<CodeInstruction> instructions)