+++ Patches/PretitlePatch.cs
@@ -14,9 +14,6 @@ namespace PeakGeneralImprovements.Patches
                 Plugin.MLS.LogMessage("Skipping pre-title screen.");
                 SceneManager.LoadScene("Title", LoadSceneMode.Single);
             }
-
-            // Late initialize things that rely on achievements
-            Plugin.LateSanitize();
         }
     }
 }